% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futres_traits.R
\name{futres_traits}
\alias{futres_traits}
\title{Access Terms From the FuTRES Ontology for Vertebrate Traits (FOVT)}
\usage{
futres_traits()
}
\value{
data.frame
}
\description{
Access traits from the FuTRES Ontology for Vertebrate Traits (FOVT)
}
\details{
The futres_traits function returns terms from the FuTRES Ontology for Vertebrate Traits (FOVT)
The response populates a data frame with: termID, label, description, and
URI.  Use the termID values in submitting termID values to the
\code{\link{futres_data}} function.  The label and description fields are
extracted from FOVT and are useful in
determining the proper term to query on.  The URI field contains a link to
the term itself which is useful for determining superclass and subclass
relationships for each term.
For more information on the FOVT ontology itself, we suggest loading the FOVT
\url{https://github.com/futres/fovt} with
protege \url{https://protege.stanford.edu/}
}
\examples{
traits <- futres_traits()
}
\keyword{lookup}
\keyword{trait}
