% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppo_data.R
\name{get_ppo_data}
\alias{get_ppo_data}
\title{Retrieve data from PPO Data portal}
\usage{
get_ppo_data(genus = NULL, specificEpithet = NULL, termID = NULL,
  fromYear = NULL, toYear = NULL, fromDay = NULL, toDay = NULL,
  bbox = NULL)
}
\arguments{
\item{genus}{a plant genus name}

\item{specificEpithet}{a plant specific epithet}

\item{termID}{a plant stage from the plant phenology ontology, e.g. obo:PPO_0002324.  Use the get_ppo_terms function in this package to get the relevant IDs for present and absent stages}

\item{fromYear}{query for years starting from this year}

\item{toYear}{query for years ending at this year}

\item{fromDay}{query for days starting from this day}

\item{toDay}{query for days ending at this day}

\item{bbox}{A lat long bounding box. Format is \code{lat,long,lat,long}. Use this website: http://boundingbox.klokantech.com/ to quickly grab a bbox (set format on bottom left to csv and be sure to switch the order from long, lat, long, lat to lat, long, lat, long)
Just set the format on the bottom left to CSV.}
}
\value{
data.frame
}
\description{
The Global Plant Phenology Data Portal (PPO data portal) is an aggregation of phenology
data from several different data sources.  Currently it contains USA-NPN,
NEON, and PEP725 data sources.  The PPO data portal harvests data using
the ppo-data-pipeline, with code available at \url{https://github.com/biocodellc/ppo-data-pipeline/}.
You may also view PPO data portal products online at \url{http://plantphenology.org/}.
}
\examples{
df <- get_ppo_data(genus = "Quercus", fromYear = 1979, toYear = 2004)
df <- get_ppo_data(bbox='44,-124,46,-122', fromDay = 1, toDay = 60)
}
\keyword{data}
\keyword{download}
