% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppo_data.R
\name{ppo_data}
\alias{ppo_data}
\title{Retrieve data from the Global Plant Phenology Ontology Data portal}
\usage{
ppo_data(genus = NULL, specificEpithet = NULL, stageID = NULL,
  fromYear = NULL, toYear = NULL, fromDay = NULL, toDay = NULL,
  bbox = NULL)
}
\arguments{
\item{genus}{a plant genus name}

\item{specificEpithet}{a plant specific epithet}

\item{stageID}{a plant stage from the plant phenology ontology, e.g. obo:PPO_0002324.  Use the ppo_stages function in this package to get the relevant IDs for present and absent stages}

\item{fromYear}{query for years starting from this year}

\item{toYear}{query for years ending at this year}

\item{fromDay}{query for days starting from this day}

\item{toDay}{query for days ending at this day}

\item{bbox}{A lat long bounding box. Format is \code{lat,long,lat,long}. Use this website: http://boundingbox.klokantech.com/ to quickly grab a bbox (set format on bottom left to csv and be sure to switch the order from long, lat, long, lat to lat, long, lat, long)
Just set the format on the bottom left to CSV.}
}
\value{
data.frame
}
\description{
Retrieve data from the Global Plant Phenology Ontology Data portal
}
\examples{
df <- ppo_data(genus = "Quercus", fromYear = 1979, toYear = 2004)
df <- ppo_data(bbox='44,-124,46,-122', fromDay = 1, toDay = 60)
}
\keyword{data}
\keyword{download}
