% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppo_get_terms.R
\name{ppo_get_terms}
\alias{ppo_get_terms}
\title{Access Terms From the Plant Phenology Ontology}
\usage{
ppo_get_terms(present = FALSE, absent = FALSE)
}
\arguments{
\item{present}{(boolean) If TRUE then return all "present" phenological stages}

\item{absent}{(boolean) IF TRUE then return all "absent" phenological stages.}
}
\value{
A data frame with the columns: termID, label, description, and URI.
}
\description{
Access present and absent terms from the Plant Phenology Ontology
}
\details{
The ppo_terms function returns terms from the Plant Phenology Ontology (PPO).
The termID or label column  can be used to query the `termID` or
`mapped_traits` to \code{\link{ppo_data}}. The label and description
columns are useful in determining the trait to query on. The URI column
contains a link to the term itself which is useful for determining superclass
and subclass relationships for each term.
}
\examples{
presentTerms <- ppo_terms(present = TRUE)
absentTerms <- ppo_terms(absent = TRUE)
allTerms <- ppo_terms(present = TRUE, absent = TRUE)
fruitTerms <- grep("fruit", ppo_filters$mapped_traits$label, value = TRUE)
ppo_data(genus = c( "Pinus", "Quercus"),
         specificEpithet = "palustris",
         mapped_traits = fruitTerms,
         fromYear = 2016, toYear = 2017)
}
\keyword{lookup}
\keyword{trait}
