% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppo_stages.R
\name{ppo_stages}
\alias{ppo_stages}
\title{Retrieve data from the Plant Phenology Ontology}
\usage{
ppo_stages(present = TRUE)
}
\arguments{
\item{present}{TRUE or FALSE. Defaults to TRUE. If TRUE then return all "present" phenological stages from the PPO.  If FALSE then return all "absent" phenological stages.}
}
\value{
data.frame
}
\description{
Retrieve data from the Plant Phenology Ontology
}
\examples{
df <- ppo_stages(present = TRUE)
}
