% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppo_traits.R
\name{ppo_traits}
\alias{ppo_traits}
\title{Get the traits data from the [ppo_data()] event ids.}
\usage{
ppo_traits(x, sorted = TRUE, flatten_traits = TRUE, flatten_all = FALSE)
}
\arguments{
\item{x}{(object) As returned from link(rppo)[ppo_data]}

\item{sorted}{(logical) Should the output traits be sorted by category. Default : TRUE}

\item{flatten_traits}{(logical) Should the traits list be melted in a data.frame. Default : TRUE}

\item{flatten_all}{(logical) Should the output list be melted in a data.frame. Default : FALSE}
}
\value{
If sorted is TRUE, a list for each event id containing a list
        with the following elements:
        \itemize{
         \item {`metadata`: A data frame containing metadata}
         \item {`taxonomy`: A data.frame containing the taxonomy}
         \item {`traits`: if flatten_traits is TRUE, a melted data.frame, else a list
                          containing the traits value}
        }
        Else, a list of data.frames for each event id.
        IF flatten_all is TRUE, the list is flatten to a data.frame
}
\description{
Get the traits data from the [ppo_data()] event ids.
}
\examples{
r1 <- ppo_data(genus = "Quercus", termID='obo:PPO_0002313', limit=10, timeLimit = 4)
r1_traits <- ppo_traits(r1)
}
