% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppo_traits.R
\name{ppo_traits_flatten}
\alias{ppo_traits_flatten}
\title{Turn the traits element or the entire list returned by [pp_traits()]
into a data.frame.}
\usage{
ppo_traits_flatten(x, flatten_all = FALSE)
}
\arguments{
\item{x}{(object) A list or data.frame as returned from link(rppo)[ppo_traits]}

\item{flatten_all}{(logical) Should the output list be melted in a data.frame. Default : FALSE}
}
\value{
@return A list for each event id containing a list
        with the following elements:
        \itemize{
         \item {`metadata`: A data frame containing metadata}
         \item {`taxonomy`: A data.frame containing the taxonomy}
         \item {`traits`: A melted data.frame containing the traits value}
        }
        IF flatten_all is TRUE, the list is flatten to a data.frame
}
\description{
Turn the traits element or the entire list returned by [pp_traits()]
into a data.frame.
}
\examples{
r1_traits <- ppo_traits(r1,  sort = FALSE, flatten_traits = FALSE)
r1_traits <- ppo_traits_flatten(r1_traits, flatten_all = TRUE)
}
