% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{ppotest}
\alias{ppotest}
\title{A Test function}
\usage{
ppotest(happy = TRUE)
}
\arguments{
\item{state}{Are you happy = true}
}
\description{
This function is just for testing
}
\examples{
cat_function()
}
\keyword{test}
