% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfutres.R
\docType{package}
\name{rfutres-package}
\alias{rfutres}
\alias{rfutres-package}
\title{Access the FuTRES Data Portal}
\description{
Access data from the futres data portal
(FUTRES data portal) and trait terms from the FuTRES Ontology for Vertebrate Traits
(FOVT)
}
\details{
\pkg{rfutres} enables users to query the FuTRES 
data portal, which is visible online at
\url{https://futres.org/}. The FOVT data portal is an aggregation
of trait data from several different data sources.  
contains USA-NPN, NEON, and PEP725 data sources.  The FuTRES data portal
harvests data using the futres-data-pipeline, with code available at
\url{https://github.com/futres/fovt-data-pipeline/}.  All trait 
terms in the data portal are aligned using the FutRES Ontology for Vertebrate Traits
(FOVT), available at \url{https://github.com/futres/fovt}.


Two functions  are contained in the \pkg{rfutres}:
\code{\link{futres_traits}} allows users to discover 
vertebrate traits while \code{\link{futres_data}} enables
users to query the FuTRES data portal. The \pkg{rfutres} package source code is
available at \url{https://github.com/ropensci/rfutres/}.
}
\author{
\strong{Maintainer}: John Deck \email{jdeck88@gmail.com}

Authors:
\itemize{
  \item John Deck \email{jdeck88@gmail.com}
  \item Meghan Balk \email{meghan.balk@gmail.com}
  \item Neeka Sewnath \email{neeka.sewnath@gmail.com}
  \item Robert Guralnick \email{rguralnick@flmnh.ufl.edu}
}

}
