% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppo.R
\docType{package}
\name{rppo-package}
\alias{rppo}
\alias{rppo-package}
\title{Access the Global Plant Phenology Data Portal}
\description{
\pkg{rppo} enables users to query the Global Plant Phenology
Data Portal (PPO data portal). The Plant Phenology Ontology is available
at \url{https://github.com/PlantPhenoOntology/ppo} and information on data
processing of PPO data is available at
\url{https://github.com/biocodellc/ppo-data-pipeline}

Two functions  are contained in the \pkg{rppo}:
\code{\link{ppo_terms}} allows users to discover present and absent
phenological stages,  \code{\link{ppo_data}} enables users to query the
PPO data portal.  Both functions return data frames.  The \pkg{rppo} package
source code is available at \url{https://github.com/biococellc/rppo}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/biocodellc/rppo}
  \item Report bugs at \url{https://github.com/biocodellc/rppo/issues}
}

}
\author{
\strong{Maintainer}: John Deck \email{jdeck88@gmail.com}

Authors:
\itemize{
  \item Brian Stucky \email{stuckyb@flmnh.ufl.edu}
  \item Ramona Walls \email{rwalls@cyverse.org}
  \item Kjell Bolmgren \email{Kjell.Bolmgren@slu.se}
  \item Ellen Denny \email{ellen@usanpn.org}
  \item Robert Guralnick \email{rguralnick@flmnh.ufl.edu}
}

}
