% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppo_data.R
\name{ppo_data}
\alias{ppo_data}
\title{Retrieve data from the Global Plant Phenology Ontology Data portal}
\usage{
ppo_data(genus = NULL, specificEpithet = NULL, trait = NULL,
  fromYear = NULL, toYear = NULL, fromDay = NULL, toDay = NULL,
  bbox = NULL, limit = NULL, offset = NULL, quiet = FALSE)
}
\arguments{
\item{genus}{a plant genus name}

\item{specificEpithet}{a plant specific epithet}

\item{trait}{a plant trait}

\item{fromYear}{query for years starting from this year}

\item{toYear}{query for years ending at this year}

\item{fromDay}{query for days starting from this day}

\item{toDay}{query for days ending at this day}

\item{bbox}{A lat long bounding box. Format is \code{lat,long,lat,long}. Use this website: http://boundingbox.klokantech.com/ to quickly grab a bbox (set format on bottom left to csv and be sure to switch the order from long, lat, long, lat to lat, long, lat, long)
Just set the format on the bottom left to CSV.}

\item{limit}{A numeric value to limit number of records}

\item{offset}{An offset best used with limit as a way to paginate records}

\item{quiet}{If true, any informative messages will be suppressed}
}
\value{
data.frame
}
\description{
Retrieve data from the Global Plant Phenology Ontology Data portal
}
\examples{
  
# data <- ppo_data(genus = "Quercus", yearFrom = 2017)
# data <- ppo_data(bbox = '37.77,-122.46,37.76,-122.47')
# fail <- aw_data(scientific_name = "auberti levithorax") # This should fail gracefully
}
\keyword{data}
\keyword{download}
