% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfutres.R
\docType{package}
\name{rfutres-package}
\alias{rfutres}
\alias{rfutres-package}
\title{Access the FuTRES Data Portal and Traits}
\description{
Access data from the FuTRES data portal
trait terms from the FuTRES Ontology for Vertebrate Traits
(FOVT)
}
\details{
\pkg{rfutres} enables users to query the futres data portal,
which is visible online at
\url{https://futres.org/}. The FuTRES data portal is an aggregation
of vertebrate trait data from several different data sources.  
The FuTRES data portal
harvests data using  the fovt-data-pipeline, with code available at
\url{https://github.com/futres/fovt-data-pipeline/}.  All trait
terms in the data portal are aligned using the FuTRES Ontology
for Vertebrate Traits (FOVT)
available at \url{https://github.com/futres/fovt}.


Two functions  are contained in the \pkg{rfutres}:
\code{\link{futres_traits}} allows users to discover 
trait terms while \code{\link{futres_data}} enables
users to query the FuTRES data portal. The \pkg{rfutres} package source code is
available at \url{https://github.com/futres/rfutres/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/futres/rfutres/}
  \item Report bugs at \url{https://github.com/futres/rfutres/issues}
}

}
\author{
\strong{Maintainer}: Meghan Balk \email{meghan.balk@gmail.com}

Authors:
\itemize{
  \item Neeka Sewnath \email{neeka.sewnath@gmail.com}
  \item John Deck \email{jdeck88@gmail.com}
  \item Robert Guralnick \email{rguralnick@flmnh.ufl.edu}
}

}
