% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppo.R
\docType{package}
\name{rppo-package}
\alias{rppo}
\alias{rppo-package}
\title{Access the Global Plant Phenology Data Portal}
\description{
\pkg{rppo} enables users to query the Global Plant Phenology Data Portal
(PPO data portal). The Plant Phenology Ontology is available at \url{https://github.com/PlantPhenoOntology/ppo}
and information on data processing of PPO data is available at \url{https://github.com/biocodellc/ppo-data-pipeline}

Two functions  are contained in the \pkg{rppo}:
\code{\link{ppo_terms}} allows users to discover present and absent
phenological stages,  \code{\link{ppo_data}} enables users to query the
PPO data portal.  Both functions return data frames.  The \pkg{rppo} package
source code is available at \url{https://github.com/biococellc/rppo}.
}
\author{
\strong{Maintainer}: John Deck \email{jdeck88@gmail.com}

}
